+++ src/custom_views/routes.js
@@ -1,1 +1,12 @@
-export const custom_views_routes = [];
+// import { CustomViewComponent } from "./components/custom_view";
+// import { CustomViewLoader } from "./api/custom_view";
+// import ErrorPage from "../error_page";
+
+export const custom_views_routes = [
+    // {
+    //     path: "/",
+    //     element: <CustomViewComponent />,
+    //     errorElement: <ErrorPage />,
+    //     loader: CustomViewLoader
+    // }
+];