+++ CHANGELOG.md
@@ -6,6 +6,15 @@ This project adheres to [Semantic Versioning](http://semver.org/) (<major>.<mino
 Backward incompatible (breaking) changes will only be introduced in major versions
 with advance notice in the Deprecations section of releases.
 
+## [0.2.0] - 2026-03-16
+
+- Changed custom views routes and directory structure
+- Changed Readonly Tabular Inlines columns
+- Changed Stacked Inlines forms columns
+- Fix object form inputs length
+- Fix sidebar overflow when many sidebar entries
+- Removed unused static assets
+
 ## [0.1.0] - 2026-03-16
 
 Initial version release.